Add map component to ModalMap. Add GIF as location symbol for proof of concept
