+++ src/app/components/img-swiper/img-swiper.component.html
@@ -5,8 +5,7 @@
     <ion-slide *ngFor="let slide of slides; let i = index" [ngSwitch]="slide.type">
       <app-offline-image *ngSwitchCase="'image'" [src]="slide.img" (click)="onImageClick(slide.img)">
       </app-offline-image>
-      <app-map-image *ngSwitchCase="'location'" [location]="slide.img"></app-map-image> 
-      <!--(click)="onLocationClick()"-->      
+      <app-map-image *ngSwitchCase="'location'" [location]="slide.img" (click)="onLocationClick()" ></app-map-image> 
       </ion-slide>
   </ion-slides>
   <app-offline-image *ngIf="isSingleImage" class="single" (click)="onImageClick(0, imgUrl[0])" [src]="imgUrl[0]">
